micro/cart-service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
mysqlclient>=2.2.0
requests>=2.31.0
django-cors-headers>=4.0.0